Rendering sidebar on mobile without any event listeners / taxing calculations / changes in height
